feat(connector): remove the sdk  support for braintree
